requirements.txt
expertai-nlapi
nltk
sklearn
wordcloud
streamlit
pandas
numpy
seaborn
matplotlib